Send only 25 heartbeats per http request and save leftovers to offline db
