+++ .gitignore
@@ -6,6 +6,8 @@
 *.dylib
 build/
 
+.idea/
+
 # Test binary, built with `go test -c`
 *.test
