--- cmd/heartbeat/params.go
@@ -1,664 +1,0 @@
-package heartbeat
-
-import (
-	"bufio"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"os"
-	"regexp"
-	"strconv"
-	"strings"
-	"time"
-
-	"github.com/mitchellh/go-homedir"
-	paramscmd "github.com/wakatime/wakatime-cli/cmd/params"
-	"github.com/wakatime/wakatime-cli/pkg/heartbeat"
-	"github.com/wakatime/wakatime-cli/pkg/log"
-	"github.com/wakatime/wakatime-cli/pkg/project"
-	"github.com/wakatime/wakatime-cli/pkg/regex"
-	"github.com/wakatime/wakatime-cli/pkg/vipertools"
-
-	"github.com/spf13/viper"
-)
-
-var (
-	// nolint
-	matchAllRegex = regexp.MustCompile(".*")
-	// nolint
-	pluginRegex = regexp.MustCompile(`(?i)([a-z\/0-9.]+\s)?(?P<editor>[a-z-]+)\-wakatime\/[0-9.]+`)
-)
-
-// Params contains heartbeat command parameters.
-type Params struct {
-	Category          heartbeat.Category
-	CursorPosition    *int
-	Entity            string
-	EntityType        heartbeat.EntityType
-	ExtraHeartbeats   []heartbeat.Heartbeat
-	IsWrite           *bool
-	Language          *string
-	LanguageAlternate string
-	LineNumber        *int
-	LinesInFile       *int
-	LocalFile         string
-	OfflineDisabled   bool
-	OfflineQueueFile  string
-	OfflineSyncMax    int
-	Time              float64
-	API               paramscmd.API
-	Filter            FilterParams
-	Project           ProjectParams
-	Sanitize          SanitizeParams
-}
-
-func (p Params) String() string {
-	var cursorPosition string
-	if p.CursorPosition != nil {
-		cursorPosition = strconv.Itoa(*p.CursorPosition)
-	}
-
-	var isWrite bool
-	if p.IsWrite != nil {
-		isWrite = *p.IsWrite
-	}
-
-	var language string
-	if p.Language != nil {
-		language = *p.Language
-	}
-
-	var lineNumber string
-	if p.LineNumber != nil {
-		lineNumber = strconv.Itoa(*p.LineNumber)
-	}
-
-	var linesInFile string
-	if p.LinesInFile != nil {
-		linesInFile = strconv.Itoa(*p.LinesInFile)
-	}
-
-	return fmt.Sprintf(
-		"category: '%s', cursor position: '%s', entity: '%s', entity type: '%s',"+
-			" num extra heartbeats: %d, is write: %t, language: '%s',"+
-			" line number: '%s', lines in file: '%s', offline disabled: %t,"+
-			" offline queue file: '%s', offline sync max: %d, time: %.5f, api params: (%s),"+
-			" filter params: (%s), project params: (%s), sanitize params: (%s)",
-		p.Category,
-		cursorPosition,
-		p.Entity,
-		p.EntityType,
-		len(p.ExtraHeartbeats),
-		isWrite,
-		language,
-		lineNumber,
-		linesInFile,
-		p.OfflineDisabled,
-		p.OfflineQueueFile,
-		p.OfflineSyncMax,
-		p.Time,
-		p.API,
-		p.Filter,
-		p.Project,
-		p.Sanitize,
-	)
-}
-
-// FilterParams contains heartbeat filtering related command parameters.
-type FilterParams struct {
-	Exclude                    []regex.Regex
-	ExcludeUnknownProject      bool
-	Include                    []regex.Regex
-	IncludeOnlyWithProjectFile bool
-}
-
-func (p FilterParams) String() string {
-	return fmt.Sprintf(
-		"exclude: '%s', exclude unknown project: %t, include: '%s', include only with project file: %t",
-		p.Exclude,
-		p.ExcludeUnknownProject,
-		p.Include,
-		p.IncludeOnlyWithProjectFile,
-	)
-}
-
-// ProjectParams params for project name sanitization.
-type ProjectParams struct {
-	Alternate        string
-	DisableSubmodule []regex.Regex
-	MapPatterns      []project.MapPattern
-	Override         string
-}
-
-func (p ProjectParams) String() string {
-	return fmt.Sprintf(
-		"alternate: '%s', disable submodule: '%s', map patterns: '%s', override: '%s'",
-		p.Alternate,
-		p.DisableSubmodule,
-		p.MapPatterns,
-		p.Override,
-	)
-}
-
-// SanitizeParams params for heartbeat sanitization.
-type SanitizeParams struct {
-	HideBranchNames  []regex.Regex
-	HideFileNames    []regex.Regex
-	HideProjectNames []regex.Regex
-}
-
-func (p SanitizeParams) String() string {
-	return fmt.Sprintf(
-		"hide branch names: '%s', hide file names: '%s', hide project names: '%s'",
-		p.HideBranchNames,
-		p.HideFileNames,
-		p.HideProjectNames,
-	)
-}
-
-// LoadParams loads heartbeat config params from viper.Viper instance. Returns ErrAuth
-// if failed to retrieve api key.
-func LoadParams(v *viper.Viper) (Params, error) {
-	return LoadParamsWithAPIKey(v, true)
-}
-
-// LoadParamsWithAPIKey loads heartbeat config params from viper.Viper instance.
-// When apiKeyRequired is true, returns ErrAuth if failed to retrieve api key.
-func LoadParamsWithAPIKey(v *viper.Viper, apiKeyRequired bool) (Params, error) {
-	params, err := paramscmd.Load(v, apiKeyRequired)
-	if err != nil {
-		return Params{}, err
-	}
-
-	var category heartbeat.Category
-
-	if categoryStr := vipertools.GetString(v, "category"); categoryStr != "" {
-		parsed, err := heartbeat.ParseCategory(categoryStr)
-		if err != nil {
-			return Params{}, fmt.Errorf("failed to parse category: %s", err)
-		}
-
-		category = parsed
-	}
-
-	var cursorPosition *int
-	if pos := v.GetInt("cursorpos"); v.IsSet("cursorpos") {
-		cursorPosition = heartbeat.Int(pos)
-	}
-
-	var entity string
-
-	entity, ok := vipertools.FirstNonEmptyString(v, "entity", "file")
-	if !ok {
-		return Params{}, errors.New("failed to retrieve entity")
-	}
-
-	entityExpanded, err := homedir.Expand(entity)
-	if err != nil {
-		return Params{}, fmt.Errorf("failed expanding entity: %s", err)
-	}
-
-	var entityType heartbeat.EntityType
-
-	if entityTypeStr := vipertools.GetString(v, "entity-type"); entityTypeStr != "" {
-		parsed, err := heartbeat.ParseEntityType(entityTypeStr)
-		if err != nil {
-			return Params{}, fmt.Errorf("failed to parse entity type: %s", err)
-		}
-
-		entityType = parsed
-	}
-
-	var extraHeartbeats []heartbeat.Heartbeat
-
-	if v.GetBool("extra-heartbeats") {
-		extraHeartbeats, err = readExtraHeartbeats()
-		if err != nil {
-			log.Errorf("failed to read extra heartbeats: %s", err)
-		}
-	}
-
-	var isWrite *bool
-	if b := v.GetBool("write"); v.IsSet("write") {
-		isWrite = heartbeat.Bool(b)
-	}
-
-	var lineNumber *int
-	if num := v.GetInt("lineno"); v.IsSet("lineno") {
-		lineNumber = heartbeat.Int(num)
-	}
-
-	var linesInFile *int
-	if num := v.GetInt("lines-in-file"); v.IsSet("lines-in-file") {
-		linesInFile = heartbeat.Int(num)
-	}
-
-	timeSecs := v.GetFloat64("time")
-	if timeSecs == 0 {
-		timeSecs = float64(time.Now().UnixNano()) / 1000000000
-	}
-
-	projectParams, err := loadProjectParams(v)
-	if err != nil {
-		return Params{}, fmt.Errorf("failed to parse project params: %s", err)
-	}
-
-	sanitizeParams, err := loadSanitizeParams(v)
-	if err != nil {
-		return Params{}, fmt.Errorf("failed to load sanitize params: %s", err)
-	}
-
-	var language *string
-	if l := vipertools.GetString(v, "language"); l != "" {
-		language = &l
-	}
-
-	return Params{
-		Category:          category,
-		CursorPosition:    cursorPosition,
-		Entity:            entityExpanded,
-		ExtraHeartbeats:   extraHeartbeats,
-		EntityType:        entityType,
-		IsWrite:           isWrite,
-		Language:          language,
-		LanguageAlternate: vipertools.GetString(v, "alternate-language"),
-		LineNumber:        lineNumber,
-		LinesInFile:       linesInFile,
-		LocalFile:         vipertools.GetString(v, "local-file"),
-		OfflineDisabled:   params.OfflineDisabled,
-		OfflineQueueFile:  params.OfflineQueueFile,
-		OfflineSyncMax:    params.OfflineSyncMax,
-		Time:              timeSecs,
-		API:               params.API,
-		Filter:            loadFilterParams(v),
-		Project:           projectParams,
-		Sanitize:          sanitizeParams,
-	}, nil
-}
-
-func readExtraHeartbeats() ([]heartbeat.Heartbeat, error) {
-	in := bufio.NewReader(os.Stdin)
-
-	input, err := in.ReadString('\n')
-	if err != nil {
-		return nil, fmt.Errorf("failed to read data from stdin: %s", err)
-	}
-
-	heartbeats, err := parseExtraHeartbeat(input)
-	if err != nil {
-		return nil, fmt.Errorf("failed to json decode: %s", err)
-	}
-
-	return heartbeats, nil
-}
-
-func parseExtraHeartbeat(data string) ([]heartbeat.Heartbeat, error) {
-	var incoming []struct {
-		Category          heartbeat.Category `json:"category"`
-		CursorPosition    interface{}        `json:"cursorpos"`
-		Entity            string             `json:"entity"`
-		EntityType        string             `json:"entity_type"`
-		Type              string             `json:"type"`
-		IsWrite           interface{}        `json:"is_write"`
-		Language          *string            `json:"language"`
-		LanguageAlternate string             `json:"alternate_language"`
-		LineNumber        interface{}        `json:"lineno"`
-		Lines             interface{}        `json:"lines"`
-		Project           string             `json:"project"`
-		ProjectAlternate  string             `json:"alternate_project"`
-		Time              interface{}        `json:"time"`
-		Timestamp         interface{}        `json:"timestamp"`
-	}
-
-	err := json.Unmarshal([]byte(data), &incoming)
-	if err != nil {
-		return nil, fmt.Errorf("failed to json decode from data %q: %s", string(data), err)
-	}
-
-	var heartbeats []heartbeat.Heartbeat
-
-	for _, h := range incoming {
-		h.Entity, err = homedir.Expand(h.Entity)
-		if err != nil {
-			return nil, fmt.Errorf("failed expanding entity: %s", err)
-		}
-
-		var entityType heartbeat.EntityType
-
-		// Both type or entity_type are acceptable here. Type takes precedence.
-		entityTypeStr := firstNonEmptyString(h.Type, h.EntityType)
-		if entityTypeStr != "" {
-			entityType, err = heartbeat.ParseEntityType(entityTypeStr)
-			if err != nil {
-				return nil, err
-			}
-		}
-
-		var cursorPosition *int
-
-		switch cursorPositionVal := h.CursorPosition.(type) {
-		case float64:
-			cursorPosition = heartbeat.Int(int(cursorPositionVal))
-		case string:
-			val, err := strconv.Atoi(cursorPositionVal)
-			if err != nil {
-				return nil, fmt.Errorf("failed to convert cursor position to int: %s", err)
-			}
-
-			cursorPosition = heartbeat.Int(val)
-		}
-
-		var isWrite *bool
-
-		switch isWriteVal := h.IsWrite.(type) {
-		case bool:
-			isWrite = heartbeat.Bool(isWriteVal)
-		case string:
-			val, err := strconv.ParseBool(isWriteVal)
-			if err != nil {
-				return nil, fmt.Errorf("failed to convert is write to bool: %s", err)
-			}
-
-			isWrite = heartbeat.Bool(val)
-		}
-
-		var lineNumber *int
-
-		switch lineNumberVal := h.LineNumber.(type) {
-		case float64:
-			lineNumber = heartbeat.Int(int(lineNumberVal))
-		case string:
-			val, err := strconv.Atoi(lineNumberVal)
-			if err != nil {
-				return nil, fmt.Errorf("failed to convert line number to int: %s", err)
-			}
-
-			lineNumber = heartbeat.Int(val)
-		}
-
-		var lines *int
-
-		switch linesVal := h.Lines.(type) {
-		case float64:
-			lines = heartbeat.Int(int(linesVal))
-		case string:
-			val, err := strconv.Atoi(linesVal)
-			if err != nil {
-				return nil, fmt.Errorf("failed to convert lines to int: %s", err)
-			}
-
-			lines = heartbeat.Int(val)
-		}
-
-		var time float64
-
-		switch timeVal := h.Time.(type) {
-		case float64:
-			time = timeVal
-		case string:
-			val, err := strconv.ParseFloat(timeVal, 64)
-			if err != nil {
-				return nil, fmt.Errorf("failed to convert time to float64: %s", err)
-			}
-
-			time = val
-		}
-
-		var timestamp float64
-
-		switch timestampVal := h.Timestamp.(type) {
-		case float64:
-			timestamp = timestampVal
-		case string:
-			val, err := strconv.ParseFloat(timestampVal, 64)
-			if err != nil {
-				return nil, fmt.Errorf("failed to convert timestamp to float64: %s", err)
-			}
-
-			timestamp = val
-		}
-
-		var timestampParsed float64
-
-		switch {
-		case h.Time != nil && h.Time != 0:
-			timestampParsed = time
-		case h.Timestamp != nil && h.Timestamp != 0:
-			timestampParsed = timestamp
-		default:
-			return nil, fmt.Errorf("skipping extra heartbeat, as no valid timestamp was defined")
-		}
-
-		heartbeats = append(heartbeats, heartbeat.Heartbeat{
-			Category:          h.Category,
-			CursorPosition:    cursorPosition,
-			Entity:            h.Entity,
-			EntityType:        entityType,
-			IsWrite:           isWrite,
-			Language:          h.Language,
-			LanguageAlternate: h.LanguageAlternate,
-			LineNumber:        lineNumber,
-			Lines:             lines,
-			ProjectAlternate:  h.ProjectAlternate,
-			ProjectOverride:   h.Project,
-			Time:              timestampParsed,
-		})
-	}
-
-	return heartbeats, nil
-}
-
-func parseEditorFromPlugin(plugin string) (string, error) {
-	match := pluginRegex.FindStringSubmatch(plugin)
-	paramsMap := make(map[string]string)
-
-	for i, name := range pluginRegex.SubexpNames() {
-		if i > 0 && i <= len(match) {
-			paramsMap[name] = match[i]
-		}
-	}
-
-	if len(paramsMap) == 0 || paramsMap["editor"] == "" {
-		return "", fmt.Errorf("plugin malformed: %s", plugin)
-	}
-
-	return paramsMap["editor"], nil
-}
-
-func loadFilterParams(v *viper.Viper) FilterParams {
-	exclude := v.GetStringSlice("exclude")
-	exclude = append(exclude, v.GetStringSlice("settings.exclude")...)
-	exclude = append(exclude, v.GetStringSlice("settings.ignore")...)
-
-	var excludePatterns []regex.Regex
-
-	for _, s := range exclude {
-		compiled, err := regex.Compile(s)
-		if err != nil {
-			log.Warnf("failed to compile exclude regex pattern %q", s)
-			continue
-		}
-
-		excludePatterns = append(excludePatterns, compiled)
-	}
-
-	include := v.GetStringSlice("include")
-	include = append(include, v.GetStringSlice("settings.include")...)
-
-	var includePatterns []regex.Regex
-
-	for _, s := range include {
-		compiled, err := regex.Compile(s)
-		if err != nil {
-			log.Warnf("failed to compile include regex pattern %q", s)
-			continue
-		}
-
-		includePatterns = append(includePatterns, compiled)
-	}
-
-	return FilterParams{
-		Exclude: excludePatterns,
-		ExcludeUnknownProject: vipertools.FirstNonEmptyBool(
-			v,
-			"exclude-unknown-project",
-			"settings.exclude_unknown_project",
-		),
-		Include: includePatterns,
-		IncludeOnlyWithProjectFile: vipertools.FirstNonEmptyBool(
-			v,
-			"include-only-with-project-file",
-			"settings.include_only_with_project_file",
-		),
-	}
-}
-
-func loadSanitizeParams(v *viper.Viper) (SanitizeParams, error) {
-	var params SanitizeParams
-
-	// hide branch names
-	hideBranchNamesStr, _ := vipertools.FirstNonEmptyString(
-		v,
-		"hide-branch-names",
-		"settings.hide_branch_names",
-		"settings.hide_branchnames",
-		"settings.hidebranchnames",
-	)
-
-	hideBranchNamesPatterns, err := parseBoolOrRegexList(hideBranchNamesStr)
-	if err != nil {
-		return SanitizeParams{}, fmt.Errorf(
-			"failed to parse regex hide branch names param %q: %s",
-			hideBranchNamesStr,
-			err,
-		)
-	}
-
-	params.HideBranchNames = hideBranchNamesPatterns
-
-	// hide project names
-	hideProjectNamesStr, _ := vipertools.FirstNonEmptyString(
-		v,
-		"hide-project-names",
-		"settings.hide_project_names",
-		"settings.hide_projectnames",
-		"settings.hideprojectnames",
-	)
-
-	hideProjectNamesPatterns, err := parseBoolOrRegexList(hideProjectNamesStr)
-	if err != nil {
-		return SanitizeParams{}, fmt.Errorf(
-			"failed to parse regex hide project names param %q: %s",
-			hideProjectNamesStr,
-			err,
-		)
-	}
-
-	params.HideProjectNames = hideProjectNamesPatterns
-
-	// hide file names
-	hideFileNamesStr, _ := vipertools.FirstNonEmptyString(
-		v,
-		"hide-file-names",
-		"hide-filenames",
-		"hidefilenames",
-		"settings.hide_file_names",
-		"settings.hide_filenames",
-		"settings.hidefilenames",
-	)
-
-	hideFileNamesPatterns, err := parseBoolOrRegexList(hideFileNamesStr)
-	if err != nil {
-		return SanitizeParams{}, fmt.Errorf(
-			"failed to parse regex hide file names param %q: %s",
-			hideFileNamesStr,
-			err,
-		)
-	}
-
-	params.HideFileNames = hideFileNamesPatterns
-
-	return params, nil
-}
-
-func loadProjectParams(v *viper.Viper) (ProjectParams, error) {
-	if v == nil {
-		return ProjectParams{}, errors.New("viper instance unset")
-	}
-
-	disableSubmodule, err := parseBoolOrRegexList(vipertools.GetString(v, "git.submodules_disabled"))
-	if err != nil {
-		return ProjectParams{}, fmt.Errorf(
-			"failed to parse regex submodules disabled param: %s",
-			err,
-		)
-	}
-
-	var mapPatterns []project.MapPattern
-
-	projectMap := v.GetStringMapString("projectmap")
-
-	for k, s := range projectMap {
-		compiled, err := regexp.Compile(k)
-		if err != nil {
-			log.Warnf("failed to compile projectmap regex pattern %q", k)
-			continue
-		}
-
-		mapPatterns = append(mapPatterns, project.MapPattern{
-			Name:  s,
-			Regex: compiled,
-		})
-	}
-
-	return ProjectParams{
-		Alternate:        vipertools.GetString(v, "alternate-project"),
-		DisableSubmodule: disableSubmodule,
-		MapPatterns:      mapPatterns,
-		Override:         vipertools.GetString(v, "project"),
-	}, nil
-}
-
-func parseBoolOrRegexList(s string) ([]regex.Regex, error) {
-	var patterns []regex.Regex
-
-	s = strings.ReplaceAll(s, "\r", "\n")
-	s = strings.Trim(s, "\n\t ")
-
-	switch {
-	case s == "":
-		break
-	case strings.ToLower(s) == "false":
-		break
-	case strings.ToLower(s) == "true":
-		patterns = []regex.Regex{matchAllRegex}
-	default:
-		splitted := strings.Split(s, "\n")
-		for _, s := range splitted {
-			s = strings.Trim(s, "\n\t ")
-			if s == "" {
-				continue
-			}
-
-			compiled, err := regex.Compile(s)
-			if err != nil {
-				return nil, fmt.Errorf("failed to compile regex %q: %s", s, err)
-			}
-
-			patterns = append(patterns, compiled)
-		}
-	}
-
-	return patterns, nil
-}
-
-// firstNonEmptyString accepts multiple values and return the first non empty string value.
-func firstNonEmptyString(values ...string) string {
-	for _, v := range values {
-		if v != "" {
-			return v
-		}
-	}
-
-	return ""
-}
