--- cmd/heartbeat/params_test.go
@@ -1,1444 +1,0 @@
-package heartbeat_test
-
-import (
-	"fmt"
-	"os"
-	"path/filepath"
-	"regexp"
-	"strings"
-	"testing"
-	"time"
-
-	cmdheartbeat "github.com/wakatime/wakatime-cli/cmd/heartbeat"
-	"github.com/wakatime/wakatime-cli/pkg/heartbeat"
-	"github.com/wakatime/wakatime-cli/pkg/project"
-	"github.com/wakatime/wakatime-cli/pkg/regex"
-	"gopkg.in/ini.v1"
-
-	"github.com/spf13/viper"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func TestLoadParams_AlternateProject(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("alternate-project", "web")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, "web", params.Project.Alternate)
-}
-
-func TestLoadParams_AlternateProject_Unset(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Empty(t, params.Project.Alternate)
-}
-
-func TestLoadParams_Category(t *testing.T) {
-	tests := map[string]heartbeat.Category{
-		"coding":         heartbeat.CodingCategory,
-		"browsing":       heartbeat.BrowsingCategory,
-		"building":       heartbeat.BuildingCategory,
-		"code reviewing": heartbeat.CodeReviewingCategory,
-		"debugging":      heartbeat.DebuggingCategory,
-		"designing":      heartbeat.DesigningCategory,
-		"indexing":       heartbeat.IndexingCategory,
-		"manual testing": heartbeat.ManualTestingCategory,
-		"running tests":  heartbeat.RunningTestsCategory,
-		"writing tests":  heartbeat.WritingTestsCategory,
-	}
-
-	for name, category := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("category", name)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, category, params.Category)
-		})
-	}
-}
-
-func TestLoadParams_Category_Default(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, heartbeat.CodingCategory, params.Category)
-}
-
-func TestLoadParams_Category_Invalid(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("category", "invalid")
-
-	_, err := cmdheartbeat.LoadParams(v)
-	require.Error(t, err)
-
-	assert.Equal(t, "failed to parse category: invalid category \"invalid\"", err.Error())
-}
-
-func TestLoadParams_CursorPosition(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("entity", "/path/to/file")
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("cursorpos", 42)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, 42, *params.CursorPosition)
-}
-
-func TestLoadParams_CursorPosition_Zero(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("entity", "/path/to/file")
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("cursorpos", 0)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, 0, *params.CursorPosition)
-}
-
-func TestLoadParams_CursorPosition_Unset(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("entity", "/path/to/file")
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Nil(t, params.CursorPosition)
-}
-
-func TestLoadParams_Entity_EntityFlagTakesPrecedence(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("file", "ignored")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, "/path/to/file", params.Entity)
-}
-
-func TestLoadParams_Entity_FileFlag(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("file", "~/path/to/file")
-
-	home, err := os.UserHomeDir()
-	require.NoError(t, err)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, filepath.Join(home, "/path/to/file"), params.Entity)
-}
-
-func TestLoadParams_Entity_Unset(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-
-	_, err := cmdheartbeat.LoadParams(v)
-	require.Error(t, err)
-
-	assert.Equal(t, "failed to retrieve entity", err.Error())
-}
-
-func TestLoadParams_EntityType(t *testing.T) {
-	tests := map[string]heartbeat.EntityType{
-		"file":   heartbeat.FileType,
-		"domain": heartbeat.DomainType,
-		"app":    heartbeat.AppType,
-	}
-
-	for name, entityType := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("entity-type", name)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, entityType, params.EntityType)
-		})
-	}
-}
-
-func TestLoadParams_EntityType_Default(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, heartbeat.FileType, params.EntityType)
-}
-
-func TestLoadParams_EntityType_Invalid(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("entity-type", "invalid")
-
-	_, err := cmdheartbeat.LoadParams(v)
-	require.Error(t, err)
-
-	assert.Equal(t, "failed to parse entity type: invalid entity type \"invalid\"", err.Error())
-}
-
-func TestLoadParams_ExtraHeartbeats(t *testing.T) {
-	r, w, err := os.Pipe()
-	require.NoError(t, err)
-
-	defer func() {
-		r.Close()
-		w.Close()
-	}()
-
-	origStdin := os.Stdin
-
-	defer func() { os.Stdin = origStdin }()
-
-	os.Stdin = r
-
-	data, err := os.ReadFile("testdata/extra_heartbeats.json")
-	require.NoError(t, err)
-
-	go func() {
-		_, err := w.Write(data)
-		require.NoError(t, err)
-
-		w.Close()
-	}()
-
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("extra-heartbeats", true)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Len(t, params.ExtraHeartbeats, 2)
-
-	assert.NotNil(t, params.ExtraHeartbeats[0].Language)
-	assert.Equal(t, heartbeat.LanguageGo.String(), *params.ExtraHeartbeats[0].Language)
-	assert.NotNil(t, params.ExtraHeartbeats[1].Language)
-	assert.Equal(t, heartbeat.LanguagePython.String(), *params.ExtraHeartbeats[1].Language)
-
-	assert.Equal(t, []heartbeat.Heartbeat{
-		{
-			Category:          heartbeat.CodingCategory,
-			CursorPosition:    heartbeat.Int(12),
-			Entity:            "testdata/main.go",
-			EntityType:        heartbeat.FileType,
-			IsWrite:           heartbeat.Bool(true),
-			LanguageAlternate: "Golang",
-			LineNumber:        heartbeat.Int(42),
-			Lines:             heartbeat.Int(45),
-			ProjectAlternate:  "billing",
-			ProjectOverride:   "wakatime-cli",
-			Time:              1585598059,
-			// tested above
-			Language: params.ExtraHeartbeats[0].Language,
-		},
-		{
-			Category:          heartbeat.DebuggingCategory,
-			Entity:            "testdata/main.py",
-			EntityType:        heartbeat.FileType,
-			IsWrite:           nil,
-			LanguageAlternate: "Py",
-			LineNumber:        nil,
-			Lines:             nil,
-			ProjectOverride:   "wakatime-cli",
-			Time:              1585598060,
-			// tested above
-			Language: params.ExtraHeartbeats[1].Language,
-		},
-	}, params.ExtraHeartbeats)
-}
-
-func TestLoadParams_ExtraHeartbeats_WithStringValues(t *testing.T) {
-	r, w, err := os.Pipe()
-	require.NoError(t, err)
-
-	defer func() {
-		r.Close()
-		w.Close()
-	}()
-
-	origStdin := os.Stdin
-
-	defer func() { os.Stdin = origStdin }()
-
-	os.Stdin = r
-
-	data, err := os.ReadFile("testdata/extra_heartbeats_with_string_values.json")
-	require.NoError(t, err)
-
-	go func() {
-		_, err := w.Write(data)
-		require.NoError(t, err)
-
-		w.Close()
-	}()
-
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("extra-heartbeats", true)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Len(t, params.ExtraHeartbeats, 2)
-
-	assert.NotNil(t, params.ExtraHeartbeats[0].Language)
-	assert.Equal(t, heartbeat.LanguageGo.String(), *params.ExtraHeartbeats[0].Language)
-	assert.NotNil(t, params.ExtraHeartbeats[1].Language)
-	assert.Equal(t, heartbeat.LanguagePython.String(), *params.ExtraHeartbeats[1].Language)
-
-	assert.Equal(t, []heartbeat.Heartbeat{
-		{
-			Category:       heartbeat.CodingCategory,
-			CursorPosition: heartbeat.Int(12),
-			Entity:         "testdata/main.go",
-			EntityType:     heartbeat.FileType,
-			IsWrite:        heartbeat.Bool(true),
-			Language:       params.ExtraHeartbeats[0].Language,
-			Lines:          heartbeat.Int(45),
-			LineNumber:     heartbeat.Int(42),
-			Time:           1585598059,
-		},
-		{
-			Category:       heartbeat.CodingCategory,
-			CursorPosition: heartbeat.Int(13),
-			Entity:         "testdata/main.go",
-			EntityType:     heartbeat.FileType,
-			IsWrite:        heartbeat.Bool(true),
-			Language:       params.ExtraHeartbeats[1].Language,
-			LineNumber:     heartbeat.Int(43),
-			Lines:          heartbeat.Int(46),
-			Time:           1585598060,
-		},
-	}, params.ExtraHeartbeats)
-}
-
-func TestLoadParams_IsWrite(t *testing.T) {
-	tests := map[string]bool{
-		"is write":    true,
-		"is no write": false,
-	}
-
-	for name, isWrite := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("write", isWrite)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, isWrite, *params.IsWrite)
-		})
-	}
-}
-
-func TestLoadParams_IsWrite_Unset(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Nil(t, params.IsWrite)
-}
-
-func TestLoadParams_Language(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("entity", "/path/to/file")
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("language", "Go")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.NotNil(t, params.Language)
-	assert.Equal(t, heartbeat.LanguageGo.String(), *params.Language)
-}
-
-func TestLoadParams_LanguageAlternate(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("entity", "/path/to/file")
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("alternate-language", "Go")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, heartbeat.LanguageGo.String(), params.LanguageAlternate)
-	assert.Nil(t, params.Language)
-}
-
-func TestLoadParams_LineNumber(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("entity", "/path/to/file")
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("lineno", 42)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, 42, *params.LineNumber)
-}
-
-func TestLoadParams_LineNumber_Zero(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("entity", "/path/to/file")
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("lineno", 0)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, 0, *params.LineNumber)
-}
-
-func TestLoadParams_LineNumber_Unset(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("entity", "/path/to/file")
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Nil(t, params.LineNumber)
-}
-
-func TestLoadParams_LocalFile(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("entity", "/path/to/file")
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("local-file", "/path/to/file")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, "/path/to/file", params.LocalFile)
-}
-
-func TestLoadParams_Plugin(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("plugin", "plugin/10.0.0")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, "plugin/10.0.0", params.API.Plugin)
-}
-
-func TestLoadParams_Plugin_Unset(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Empty(t, params.API.Plugin)
-}
-
-func TestLoadParams_Project(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("project", "billing")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, "billing", params.Project.Override)
-}
-
-func TestLoadParams_Project_Unset(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Empty(t, params.Project.Override)
-}
-
-func TestLoadParams_ProjectMap(t *testing.T) {
-	tests := map[string]struct {
-		Entity   string
-		Regex    regex.Regex
-		Project  string
-		Expected []project.MapPattern
-	}{
-		"simple regex": {
-			Entity:  "/home/user/projects/foo/file",
-			Regex:   regexp.MustCompile("projects/foo"),
-			Project: "My Awesome Project",
-			Expected: []project.MapPattern{
-				{
-					Name:  "My Awesome Project",
-					Regex: regexp.MustCompile("projects/foo"),
-				},
-			},
-		},
-		"regex with group replacement": {
-			Entity:  "/home/user/projects/bar123/file",
-			Regex:   regexp.MustCompile(`^/home/user/projects/bar(\\d+)/`),
-			Project: "project{0}",
-			Expected: []project.MapPattern{
-				{
-					Name:  "project{0}",
-					Regex: regexp.MustCompile(`^/home/user/projects/bar(\\d+)/`),
-				},
-			},
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", test.Entity)
-			v.Set(fmt.Sprintf("projectmap.%s", test.Regex.String()), test.Project)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, test.Expected, params.Project.MapPatterns)
-		})
-	}
-}
-
-func TestLoadParams_Timeout_FlagTakesPreceedence(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("timeout", 5)
-	v.Set("settings.timeout", 10)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, 5*time.Second, params.API.Timeout)
-}
-
-func TestLoadParams_Timeout_FromConfig(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("entity", "/path/to/file")
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("settings.timeout", 10)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, 10*time.Second, params.API.Timeout)
-}
-
-func TestLoadParams_Time(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("time", 1590609206.1)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, 1590609206.1, params.Time)
-}
-
-func TestLoadParams_Time_Default(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	now := float64(time.Now().UnixNano()) / 1000000000
-	assert.GreaterOrEqual(t, now, params.Time)
-	assert.GreaterOrEqual(t, params.Time, now-60)
-}
-
-func TestLoadParams_Filter_Exclude(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("exclude", []string{".*", "wakatime.*"})
-	v.Set("settings.exclude", []string{".+", "wakatime.+"})
-	v.Set("settings.ignore", []string{".?", "wakatime.?"})
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	require.Len(t, params.Filter.Exclude, 6)
-	assert.Equal(t, ".*", params.Filter.Exclude[0].String())
-	assert.Equal(t, "wakatime.*", params.Filter.Exclude[1].String())
-	assert.Equal(t, ".+", params.Filter.Exclude[2].String())
-	assert.Equal(t, "wakatime.+", params.Filter.Exclude[3].String())
-	assert.Equal(t, ".?", params.Filter.Exclude[4].String())
-	assert.Equal(t, "wakatime.?", params.Filter.Exclude[5].String())
-}
-
-func TestLoadParams_Filter_Exclude_Multiline(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("settings.ignore", "\t.?\n\twakatime.? \t\n")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	require.Len(t, params.Filter.Exclude, 2)
-	assert.Equal(t, ".?", params.Filter.Exclude[0].String())
-	assert.Equal(t, "wakatime.?", params.Filter.Exclude[1].String())
-}
-
-func TestLoadParams_Filter_Exclude_IgnoresInvalidRegex(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("exclude", []string{".*", "["})
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	require.Len(t, params.Filter.Exclude, 1)
-	assert.Equal(t, ".*", params.Filter.Exclude[0].String())
-}
-
-func TestLoadParams_Filter_Exclude_PerlRegexPatterns(t *testing.T) {
-	tests := map[string]string{
-		"negative lookahead": `^/var/(?!www/).*`,
-		"positive lookahead": `^/var/(?=www/).*`,
-	}
-
-	for name, pattern := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("exclude", []string{pattern})
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			require.Len(t, params.Filter.Exclude, 1)
-			assert.Equal(t, pattern, params.Filter.Exclude[0].String())
-		})
-	}
-}
-
-func TestLoadParams_Filter_ExcludeUnknownProject(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("exclude-unknown-project", true)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, true, params.Filter.ExcludeUnknownProject)
-}
-
-func TestLoadParams_Filter_ExcludeUnknownProject_FromConfig(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("exclude-unknown-project", false)
-	v.Set("settings.exclude_unknown_project", true)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, true, params.Filter.ExcludeUnknownProject)
-}
-
-func TestLoadParams_Filter_Include(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("include", []string{".*", "wakatime.*"})
-	v.Set("settings.include", []string{".+", "wakatime.+"})
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	require.Len(t, params.Filter.Include, 4)
-	assert.Equal(t, ".*", params.Filter.Include[0].String())
-	assert.Equal(t, "wakatime.*", params.Filter.Include[1].String())
-	assert.Equal(t, ".+", params.Filter.Include[2].String())
-	assert.Equal(t, "wakatime.+", params.Filter.Include[3].String())
-}
-
-func TestLoadParams_Filter_Include_IgnoresInvalidRegex(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("include", []string{".*", "["})
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	require.Len(t, params.Filter.Include, 1)
-	assert.Equal(t, ".*", params.Filter.Include[0].String())
-}
-
-func TestLoadParams_Filter_Include_PerlRegexPatterns(t *testing.T) {
-	tests := map[string]string{
-		"negative lookahead": `^/var/(?!www/).*`,
-		"positive lookahead": `^/var/(?=www/).*`,
-	}
-
-	for name, pattern := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("include", []string{pattern})
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			require.Len(t, params.Filter.Include, 1)
-			assert.Equal(t, pattern, params.Filter.Include[0].String())
-		})
-	}
-}
-
-func TestLoadParams_Filter_IncludeOnlyWithProjectFile(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("include-only-with-project-file", true)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, true, params.Filter.IncludeOnlyWithProjectFile)
-}
-
-func TestLoadParams_Filter_IncludeOnlyWithProjectFile_FromConfig(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("include-only-with-project-file", false)
-	v.Set("settings.include_only_with_project_file", true)
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, true, params.Filter.IncludeOnlyWithProjectFile)
-}
-
-func TestLoadParams_SanitizeParams_HideBranchNames_True(t *testing.T) {
-	tests := map[string]string{
-		"lowercase":       "true",
-		"uppercase":       "TRUE",
-		"first uppercase": "True",
-	}
-
-	for name, viperValue := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("hide-branch-names", viperValue)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, cmdheartbeat.SanitizeParams{
-				HideBranchNames: []regex.Regex{regex.MustCompile(".*")},
-			}, params.Sanitize)
-		})
-	}
-}
-
-func TestLoadParams_SanitizeParams_HideBranchNames_False(t *testing.T) {
-	tests := map[string]string{
-		"lowercase":       "false",
-		"uppercase":       "FALSE",
-		"first uppercase": "False",
-	}
-
-	for name, viperValue := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("hide-branch-names", viperValue)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, cmdheartbeat.SanitizeParams{}, params.Sanitize)
-		})
-	}
-}
-
-func TestLoadParams_SanitizeParams_HideBranchNames_List(t *testing.T) {
-	tests := map[string]struct {
-		ViperValue string
-		Expected   []regex.Regex
-	}{
-		"regex": {
-			ViperValue: "fix.*",
-			Expected: []regex.Regex{
-				regexp.MustCompile("fix.*"),
-			},
-		},
-		"regex list": {
-			ViperValue: ".*secret.*\nfix.*",
-			Expected: []regex.Regex{
-				regexp.MustCompile(".*secret.*"),
-				regexp.MustCompile("fix.*"),
-			},
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("hide-branch-names", test.ViperValue)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, cmdheartbeat.SanitizeParams{
-				HideBranchNames: test.Expected,
-			}, params.Sanitize)
-		})
-	}
-}
-
-func TestLoadParams_SanitizeParams_HideBranchNames_FlagTakesPrecedence(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("hide-branch-names", "true")
-	v.Set("settings.hide_branch_names", "ignored")
-	v.Set("settings.hide_branchnames", "ignored")
-	v.Set("settings.hidebranchnames", "ignored")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideBranchNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideBranchNames_ConfigTakesPrecedence(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("settings.hide_branch_names", "true")
-	v.Set("settings.hide_branchnames", "ignored")
-	v.Set("settings.hidebranchnames", "ignored")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideBranchNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideBranchNames_ConfigDeprecatedOneTakesPrecedence(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("settings.hide_branchnames", "true")
-	v.Set("settings.hidebranchnames", "ignored")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideBranchNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideBranchNames_ConfigDeprecatedTwo(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("settings.hidebranchnames", "true")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideBranchNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideBranchNames_InvalidRegex(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("hide-branch-names", ".*secret.*\n[0-9+")
-
-	_, err := cmdheartbeat.LoadParams(v)
-	require.Error(t, err)
-
-	assert.True(t, strings.HasPrefix(
-		err.Error(),
-		"failed to load sanitize params:"+
-			" failed to parse regex hide branch names param \".*secret.*\\n[0-9+\":"+
-			" failed to compile regex \"[0-9+\":",
-	))
-}
-
-func TestLoadParams_SanitizeParams_HideProjectNames_True(t *testing.T) {
-	tests := map[string]string{
-		"lowercase":       "true",
-		"uppercase":       "TRUE",
-		"first uppercase": "True",
-	}
-
-	for name, viperValue := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("hide-project-names", viperValue)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, cmdheartbeat.SanitizeParams{
-				HideProjectNames: []regex.Regex{regexp.MustCompile(".*")},
-			}, params.Sanitize)
-		})
-	}
-}
-
-func TestLoadParams_SanitizeParams_HideProjectNames_False(t *testing.T) {
-	tests := map[string]string{
-		"lowercase":       "false",
-		"uppercase":       "FALSE",
-		"first uppercase": "False",
-	}
-
-	for name, viperValue := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("hide-project-names", viperValue)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, cmdheartbeat.SanitizeParams{}, params.Sanitize)
-		})
-	}
-}
-
-func TestLoadParams_SanitizeParams_HideProjecthNames_List(t *testing.T) {
-	tests := map[string]struct {
-		ViperValue string
-		Expected   []regex.Regex
-	}{
-		"regex": {
-			ViperValue: "fix.*",
-			Expected: []regex.Regex{
-				regexp.MustCompile("fix.*"),
-			},
-		},
-		"regex list": {
-			ViperValue: ".*secret.*\nfix.*",
-			Expected: []regex.Regex{
-				regexp.MustCompile(".*secret.*"),
-				regexp.MustCompile("fix.*"),
-			},
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("hide-project-names", test.ViperValue)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, cmdheartbeat.SanitizeParams{
-				HideProjectNames: test.Expected,
-			}, params.Sanitize)
-		})
-	}
-}
-
-func TestLoadParams_SanitizeParams_HideProjectNames_FlagTakesPrecedence(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("hide-project-names", "true")
-	v.Set("settings.hide_project_names", "ignored")
-	v.Set("settings.hide_projectnames", "ignored")
-	v.Set("settings.hideprojectnames", "ignored")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideProjectNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideProjectNames_ConfigTakesPrecedence(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("settings.hide_project_names", "true")
-	v.Set("settings.hide_projectnames", "ignored")
-	v.Set("settings.hideprojectnames", "ignored")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideProjectNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideProjectNames_ConfigDeprecatedOneTakesPrecedence(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("settings.hide_projectnames", "true")
-	v.Set("settings.hideprojectnames", "ignored")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideProjectNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideProjectNames_ConfigDeprecatedTwo(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("settings.hideprojectnames", "true")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideProjectNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideProjectNames_InvalidRegex(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("hide-project-names", ".*secret.*\n[0-9+")
-
-	_, err := cmdheartbeat.LoadParams(v)
-	require.Error(t, err)
-
-	assert.True(t, strings.HasPrefix(
-		err.Error(),
-		"failed to load sanitize params:"+
-			" failed to parse regex hide project names param \".*secret.*\\n[0-9+\":"+
-			" failed to compile regex \"[0-9+\":",
-	))
-}
-
-func TestLoadParams_SanitizeParams_HideFileNames_True(t *testing.T) {
-	tests := map[string]string{
-		"lowercase":       "true",
-		"uppercase":       "TRUE",
-		"first uppercase": "True",
-	}
-
-	for name, viperValue := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("hide-file-names", viperValue)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, cmdheartbeat.SanitizeParams{
-				HideFileNames: []regex.Regex{regexp.MustCompile(".*")},
-			}, params.Sanitize)
-		})
-	}
-}
-
-func TestLoadParams_SanitizeParams_HideFileNames_False(t *testing.T) {
-	tests := map[string]string{
-		"lowercase":       "false",
-		"uppercase":       "FALSE",
-		"first uppercase": "False",
-	}
-
-	for name, viperValue := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("hide-file-names", viperValue)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, cmdheartbeat.SanitizeParams{}, params.Sanitize)
-		})
-	}
-}
-
-func TestLoadParams_SanitizeParams_HideFilehNames_List(t *testing.T) {
-	tests := map[string]struct {
-		ViperValue string
-		Expected   []regex.Regex
-	}{
-		"regex": {
-			ViperValue: "fix.*",
-			Expected: []regex.Regex{
-				regexp.MustCompile("fix.*"),
-			},
-		},
-		"regex list": {
-			ViperValue: ".*secret.*\nfix.*",
-			Expected: []regex.Regex{
-				regexp.MustCompile(".*secret.*"),
-				regexp.MustCompile("fix.*"),
-			},
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("hide-file-names", test.ViperValue)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, cmdheartbeat.SanitizeParams{
-				HideFileNames: test.Expected,
-			}, params.Sanitize)
-		})
-	}
-}
-
-func TestLoadParams_SanitizeParams_HideFileNames_FlagTakesPrecedence(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("hide-file-names", "true")
-	v.Set("hide-filenames", "ignored")
-	v.Set("hidefilenames", "ignored")
-	v.Set("settings.hide_file_names", "ignored")
-	v.Set("settings.hide_filenames", "ignored")
-	v.Set("settings.hidefilenames", "ignored")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideFileNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideFileNames_FlagDeprecatedOneTakesPrecedence(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("hide-filenames", "true")
-	v.Set("hidefilenames", "ignored")
-	v.Set("settings.hide_file_names", "ignored")
-	v.Set("settings.hide_filenames", "ignored")
-	v.Set("settings.hidefilenames", "ignored")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideFileNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideFileNames_FlagDeprecatedTwoTakesPrecedence(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("hidefilenames", "true")
-	v.Set("settings.hide_file_names", "ignored")
-	v.Set("settings.hide_filenames", "ignored")
-	v.Set("settings.hidefilenames", "ignored")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideFileNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideFileNames_ConfigTakesPrecedence(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("settings.hide_file_names", "true")
-	v.Set("settings.hide_filenames", "ignored")
-	v.Set("settings.hidefilenames", "ignored")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideFileNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideFileNames_ConfigDeprecatedOneTakesPrecedence(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("settings.hide_filenames", "true")
-	v.Set("settings.hidefilenames", "ignored")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideFileNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideFileNames_ConfigDeprecatedTwo(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("settings.hidefilenames", "true")
-
-	params, err := cmdheartbeat.LoadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, cmdheartbeat.SanitizeParams{
-		HideFileNames: []regex.Regex{regexp.MustCompile(".*")},
-	}, params.Sanitize)
-}
-
-func TestLoadParams_SanitizeParams_HideFileNames_InvalidRegex(t *testing.T) {
-	v := viper.New()
-	v.SetDefault("sync-offline-activity", 1000)
-	v.Set("key", "00000000-0000-4000-8000-000000000000")
-	v.Set("entity", "/path/to/file")
-	v.Set("hide-file-names", ".*secret.*\n[0-9+")
-
-	_, err := cmdheartbeat.LoadParams(v)
-	require.Error(t, err)
-
-	assert.True(t, strings.HasPrefix(
-		err.Error(),
-		"failed to load sanitize params:"+
-			" failed to parse regex hide file names param \".*secret.*\\n[0-9+\":"+
-			" failed to compile regex \"[0-9+\":",
-	))
-}
-
-func TestLoadParams_DisableSubmodule_True(t *testing.T) {
-	tests := map[string]string{
-		"lowercase":       "true",
-		"uppercase":       "TRUE",
-		"first uppercase": "True",
-	}
-
-	for name, viperValue := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("git.submodules_disabled", viperValue)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, []regex.Regex{regexp.MustCompile(".*")}, params.Project.DisableSubmodule)
-		})
-	}
-}
-
-func TestLoadParams_DisableSubmodule_False(t *testing.T) {
-	tests := map[string]string{
-		"lowercase":       "false",
-		"uppercase":       "FALSE",
-		"first uppercase": "False",
-	}
-
-	for name, viperValue := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("git.submodules_disabled", viperValue)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, []regex.Regex(nil), params.Project.DisableSubmodule)
-		})
-	}
-}
-
-func TestLoadParams_DisableSubmodule_List(t *testing.T) {
-	tests := map[string]struct {
-		ViperValue string
-		Expected   []regex.Regex
-	}{
-		"regex": {
-			ViperValue: "fix.*",
-			Expected: []regex.Regex{
-				regexp.MustCompile("fix.*"),
-			},
-		},
-		"regex_list": {
-			ViperValue: "\n.*secret.*\nfix.*",
-			Expected: []regex.Regex{
-				regexp.MustCompile(".*secret.*"),
-				regexp.MustCompile("fix.*"),
-			},
-		},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			multilineOption := viper.IniLoadOptions(ini.LoadOptions{AllowPythonMultilineValues: true})
-			v := viper.NewWithOptions(multilineOption)
-			v.SetDefault("sync-offline-activity", 1000)
-			v.Set("key", "00000000-0000-4000-8000-000000000000")
-			v.Set("entity", "/path/to/file")
-			v.Set("git.submodules_disabled", test.ViperValue)
-
-			params, err := cmdheartbeat.LoadParams(v)
-			require.NoError(t, err)
-
-			assert.Equal(t, test.Expected, params.Project.DisableSubmodule)
-		})
-	}
-}
