+++ cmd/params/params_internal_test.go
@@ -1,4 +1,4 @@
-package heartbeat
+package params
 
 import (
 	"regexp"
