+++ cmd/params/testdata/main.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+# vim: set filetype=python
+
+from __future__ import print_statement
+
+import os, sys
+from flask import session, render_template
+import simplejson as json
+
+
+class MyClass(object):
+    """this class
+    """
+
+    def method1(self):
+        a = 1 + 2
+        b = 'hello world!'
+        for x in y:
+            print(x)
+        raise Exception()
