+++ pkg/offline/offline_test.go
@@ -454,7 +454,7 @@ func TestSync_MultipleRequests(t *testing.T) {
 	dataGo, err := os.ReadFile("testdata/heartbeat_go.json")
 	require.NoError(t, err)
 
-	for i := 0; i < 11; i++ {
+	for i := 0; i < 26; i++ {
 		insertHeartbeatRecord(t, db, "heartbeats", heartbeatRecord{
 			ID:        strconv.Itoa(i) + "1592868367.219124-file-coding-wakatime-cli-heartbeat-/tmp/main.go-true",
 			Heartbeat: string(dataGo),
@@ -473,7 +473,7 @@ func TestSync_MultipleRequests(t *testing.T) {
 
 		// first request
 		if numCalls == 1 {
-			assert.Len(t, hh, 10)
+			assert.Len(t, hh, 25)
 
 			result := heartbeat.Result{
 				Status:    http.StatusCreated,
@@ -483,7 +483,10 @@ func TestSync_MultipleRequests(t *testing.T) {
 			return []heartbeat.Result{
 				result, result, result, result, result,
 				result, result, result, result, result,
-				result,
+				result, result, result, result, result,
+				result, result, result, result, result,
+				result, result, result, result, result,
+				result, result,
 			}, nil
 		}
